correction de la gestion de la récupération de la classification

+++ class/Authority.class.php
@@ -312,6 +312,14 @@ class Authority extends DataObject
             return false;
         }
 
+        $sql = "DELETE FROM helios_retour WHERE authority_id=?";
+
+        if (! $this->db->exec($sql, [$id])) {
+            $this->errorMsg = "Erreur lors de la suppression des PES Retour.";
+            $this->db->rollback();
+            return false;
+        }
+
         $sql = "DELETE FROM authority_siret WHERE authority_id=?";
 
         if (! $this->db->exec($sql, [$id])) {
